+++ docs/_nuxt/e4c58ad.js
@@ -1,0 +1,1 @@
+(window.webpackJsonp=window.webpackJsonp||[]).push([[30],{646:function(e,n,t){"use strict";t.r(n);var l={name:"OpenApiTableNested"},c=t(1),component=Object(c.a)(l,(function(){return(0,this._self._c)("div",[this._t("default")],2)}),[],!1,null,null,null);n.default=component.exports}}]);