+++ playground2/nuxt.config.js
@@ -23,7 +23,7 @@ export default {
 
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
   plugins: [
-    { src: '@/plugins/auth' },
+    { src: '@/plugins/content' },
   ],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
@@ -36,6 +36,10 @@ export default {
 
   router: {
     base: isDev ? '/' : '/nuxt-openapi-docs-module'
+  },
+
+  generate: {
+    fallback: true
   },
 
   // Modules: https://go.nuxtjs.dev/config-modules
@@ -51,17 +55,23 @@ export default {
             return {
               'petstore_extended': 'Petstore Extended Api',
               'localization': 'Localization',
+              'page': 'Custom Page',
               'no-access': 'no access',
               'test-yaml.yaml': 'yaml ext',
               'test-json.json': 'json ext',
-              "oneof": "oneof",
-              'https://petstore.swagger.io/v2/swagger.json': 'petstore url',
+              "external-refs.yaml": "external-refs.yaml",
+              "external-refs2.yaml": "external-refs2.yaml",
+              // 'https://petstore.swagger.io/v2/swagger.json': 'petstore url',
             }
           }
            return {
              'petstore_extended': 'Petstore Extended Api',
              'localization': 'Localization',
              'no-access': 'no access',
+             'page': 'Custom Page',
+             'webhook': 'webhook',
+             "external-refs": "external-refs",
+             "external-refs2": "external-refs2",
 
              "additional-props": "additional-props",
              "allof": "allof",